+++ inlang/source-code/cli/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @inlang/cli
 
+## 1.6.0
+
+### Minor Changes
+
+- 25fe8502: refactor: remove plugin.meta and messageLintRule.meta nesting
+
+## 1.5.0
+
+### Minor Changes
+
+- dcb20e90: improve: module init got new templates with testing and more
+
 ## 1.4.0
 
 ### Minor Changes